Extract LLIL evaluation helpers
