+++ sc62015/pysc62015/emulator.py
@@ -1,7 +1,5 @@
-from typing import Dict, Set, Callable, Optional, Any, cast, Tuple, TypedDict
+from typing import Dict, Set, Optional, Any, cast, Tuple
 import enum
-from dataclasses import dataclass
-
 from .coding import FetchDecoder
 from .constants import PC_MASK
 
@@ -15,8 +13,13 @@ from binja_helpers.mock_llil import (
     MockLLIL,
     MockLabel,
     MockIfExpr,
-    MockIntrinsic,
     MockGoto,
+)
+from binja_helpers.eval_llil import (
+    Memory,
+    State,
+    ResultFlags,
+    evaluate_llil,
 )
 from binaryninja import (
     InstructionInfo,
@@ -157,110 +160,12 @@ class Registers:
                     f"Attempted to set unknown or non-base register: {reg}"
                 )
 
+    def get_by_name(self, name: str) -> int:
+        return self.get(RegisterName[name])
 
-ReadMemType = Callable[[int], int]
-WriteMemType = Callable[[int, int], None]
+    def set_by_name(self, name: str, value: int) -> None:
+        self.set(RegisterName[name], value)
 
-
-class Memory:
-    def __init__(self, read_mem: ReadMemType, write_mem: WriteMemType) -> None:
-        self.read_mem = read_mem
-        self.write_mem = write_mem
-
-    def read_byte(self, address: int) -> int:
-        return self.read_mem(address)
-
-    def write_byte(self, address: int, value: int) -> None:
-        assert 0 <= value < 256, "Value must be a byte (0-255)"
-        self.write_mem(address, value & 0xFF)
-
-    def read_bytes(self, address: int, size: int) -> int:
-        assert 0 < size <= 3, "Size must be between 1 and 3 bytes"
-        value = 0
-        for i in range(size):
-            value |= self.read_byte(address + i) << (i * 8)
-        return value
-
-    def write_bytes(self, size: int, address: int, value: int) -> None:
-        assert 0 < size <= 3
-        for i in range(size):
-            byte_value = (value >> (i * 8)) & 0xFF
-            self.write_byte(address + i, byte_value)
-
-
-@dataclass
-class State:
-    halted: bool = False
-
-
-class ResultFlags(TypedDict, total=False):
-    C: Optional[int]
-    Z: Optional[int]
-
-
-EvalLLILType = Callable[
-    [MockLLIL, Optional[int], Registers, Memory, State],
-    Tuple[Optional[int], Optional[ResultFlags]],
-]
-
-
-def evaluate_llil(
-    llil: MockLLIL, regs: Registers, memory: Memory, state: State
-) -> Tuple[Optional[int], Optional[ResultFlags]]:
-    op_name_bare = llil.bare_op()
-    llil_flags_spec = llil.flags()  # e.g., "CZ", "Z", or None
-    size = llil.width()
-    current_op_name_for_eval = op_name_bare  # Will be updated for intrinsics
-
-    if isinstance(llil, MockIntrinsic):
-        intrinsic = cast(MockIntrinsic, llil)
-        current_op_name_for_eval = f"INTRINSIC_{intrinsic.name}"
-
-    f = EVAL_LLIL.get(current_op_name_for_eval)
-    if f is None:
-        raise NotImplementedError(
-            f"Eval for {current_op_name_for_eval} not implemented"
-        )
-
-    result_value, op_defined_flags = f(llil, size, regs, memory, state)
-
-    if llil_flags_spec is not None and llil_flags_spec != "0":
-        if "Z" in llil_flags_spec:
-            # We must set FZ based on either op_defined_flags or result_value
-            fz_val_to_set: Optional[int] = None
-            if op_defined_flags:
-                fz_val_to_set = op_defined_flags.get("Z")
-
-            if fz_val_to_set is not None:
-                regs.set(RegisterName.FZ, fz_val_to_set)
-            elif isinstance(result_value, int):
-                assert (
-                    size is not None
-                ), f"FZ flag setting from result_value requires size for {current_op_name_for_eval}"
-                zero_mask = (1 << (size * 8)) - 1
-                regs.set(RegisterName.FZ, int((result_value & zero_mask) == 0))
-            # If neither, FZ is not set by this explicit mechanism (might be implicitly correct or an oversight in arch spec)
-
-        if "C" in llil_flags_spec:
-            # We must set FC based on either op_defined_flags or result_value
-            fc_val_to_set: Optional[int] = None
-            if op_defined_flags:
-                fc_val_to_set = op_defined_flags.get("C")
-
-            if fc_val_to_set is not None:
-                regs.set(RegisterName.FC, fc_val_to_set)
-            elif isinstance(result_value, int):
-                assert (
-                    size is not None
-                ), f"FC flag setting from result_value requires size for {current_op_name_for_eval}"
-                unsigned_max_for_size = (1 << (size * 8)) - 1
-                carry_flag_val = 0
-                if result_value > unsigned_max_for_size:  # Unsigned overflow
-                    carry_flag_val = 1
-                regs.set(RegisterName.FC, carry_flag_val)
-            # If neither, FC is not set by this explicit mechanism.
-
-    return result_value, op_defined_flags
 
 
 class Emulator:
@@ -336,489 +241,3 @@ class Emulator:
     def evaluate(self, llil: MockLLIL) -> Tuple[Optional[int], Optional[ResultFlags]]:
         return evaluate_llil(llil, self.regs, self.memory, self.state)
 
-
-def eval_const(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[int, Optional[ResultFlags]]:
-    result = llil.ops[0]
-    assert isinstance(result, int)
-    return result, None
-
-
-def eval_const_ptr(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[int, Optional[ResultFlags]]:
-    result = llil.ops[0]
-    assert isinstance(result, int)
-    return result, None
-
-
-def eval_reg(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[int, Optional[ResultFlags]]:
-    reg_name_enum = RegisterName(llil.ops[0].name)
-    return regs.get(reg_name_enum), None
-
-
-def eval_set_reg(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    reg_name_enum = RegisterName(llil.ops[0].name)
-    # Ensure llil.ops[1] is MockLLIL for eval
-    assert isinstance(
-        llil.ops[1], MockLLIL
-    ), "Source operand for SET_REG must be MockLLIL"
-    value_to_set, _ = evaluate_llil(llil.ops[1], regs, memory, state)
-    assert value_to_set is not None, "Value for SET_REG cannot be None"
-    regs.set(reg_name_enum, value_to_set)
-    return None, None
-
-
-def eval_flag(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[int, Optional[ResultFlags]]:
-    flag_name_enum = RegisterName(f"F{llil.ops[0].name}")
-    return regs.get(flag_name_enum), None
-
-
-def eval_set_flag(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    flag_name_enum = RegisterName(f"F{llil.ops[0].name}")
-    assert isinstance(
-        llil.ops[1], MockLLIL
-    ), "Source operand for SET_FLAG must be MockLLIL"
-    value_to_set, _ = evaluate_llil(llil.ops[1], regs, memory, state)
-    assert value_to_set is not None, "Value for SET_FLAG cannot be None"
-    regs.set(flag_name_enum, value_to_set != 0)
-    return None, None
-
-
-def _create_logical_eval(op_func: Callable[[int, int], int]) -> EvalLLILType:
-    """Factory for simple logical operations."""
-
-    def _eval(
-        llil: MockLLIL,
-        size: Optional[int],
-        regs: Registers,
-        memory: Memory,
-        state: State,
-    ) -> Tuple[int, Optional[ResultFlags]]:
-        assert isinstance(llil.ops[0], MockLLIL) and isinstance(llil.ops[1], MockLLIL)
-        op1_val, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-        op2_val, _ = evaluate_llil(llil.ops[1], regs, memory, state)
-        assert op1_val is not None and op2_val is not None
-        result = op_func(int(op1_val), int(op2_val))
-        return result, {"Z": 1 if result == 0 else 0, "C": 0}
-
-    return _eval
-
-
-def _create_arithmetic_eval(
-    op_func: Callable[[int, int], int],
-    carry_func: Callable[[int, int, int, int], int],
-) -> EvalLLILType:
-    """Factory for arithmetic operations like ADD and SUB."""
-
-    def _eval(
-        llil: MockLLIL,
-        size: Optional[int],
-        regs: Registers,
-        memory: Memory,
-        state: State,
-    ) -> Tuple[int, Optional[ResultFlags]]:
-        assert size is not None
-        assert isinstance(llil.ops[0], MockLLIL) and isinstance(llil.ops[1], MockLLIL)
-        op1_val, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-        op2_val, _ = evaluate_llil(llil.ops[1], regs, memory, state)
-        assert op1_val is not None and op2_val is not None
-
-        result_full = op_func(int(op1_val), int(op2_val))
-
-        width_bits = size * 8
-        mask = (1 << width_bits) - 1
-        result_masked = result_full & mask
-
-        flag_z = 1 if result_masked == 0 else 0
-        flag_c = carry_func(int(op1_val), int(op2_val), result_full, mask)
-
-        return result_masked, {"C": flag_c, "Z": flag_z}
-
-    return _eval
-
-
-def _create_shift_eval(op_func: Callable[..., Tuple[int, ResultFlags]]) -> EvalLLILType:
-    """Factory for shift and rotate operations."""
-
-    def _eval(
-        llil: MockLLIL,
-        size: Optional[int],
-        regs: Registers,
-        memory: Memory,
-        state: State,
-    ) -> Tuple[int, Optional[ResultFlags]]:
-        assert size is not None
-        values = []
-        for operand in llil.ops:
-            assert isinstance(operand, MockLLIL)
-            val, _ = evaluate_llil(operand, regs, memory, state)
-            assert val is not None
-            values.append(int(val))
-
-        result, flags = op_func(size, *values)
-        return result, flags
-
-    return _eval
-
-
-
-
-
-def eval_pop(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[int, Optional[ResultFlags]]:
-    assert size
-    addr = regs.get(RegisterName.S)
-    result = memory.read_bytes(addr, size)
-    regs.set(RegisterName.S, addr + size)
-    return result, None
-
-
-def eval_push(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    assert size
-    assert isinstance(llil.ops[0], MockLLIL)
-    value_to_push, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-    assert value_to_push is not None
-    addr = regs.get(RegisterName.S) - size
-    memory.write_bytes(size, addr, value_to_push)
-    regs.set(RegisterName.S, addr)
-    return None, None
-
-
-def eval_nop(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    return None, None
-
-
-def eval_unimpl(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    raise NotImplementedError(f"Low-level IL operation {llil.op} is not implemented")
-
-
-def eval_store(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    assert size
-    assert isinstance(llil.ops[0], MockLLIL) and isinstance(llil.ops[1], MockLLIL)
-    dest_addr, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-    value_to_store, _ = evaluate_llil(llil.ops[1], regs, memory, state)
-    assert dest_addr is not None and value_to_store is not None
-    memory.write_bytes(size, dest_addr, value_to_store)
-    return None, None
-
-
-def eval_load(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[int, Optional[ResultFlags]]:
-    assert size
-    assert isinstance(llil.ops[0], MockLLIL)
-    addr, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-    assert addr is not None
-    return memory.read_bytes(addr, size), None
-
-
-def eval_ret(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    assert isinstance(llil.ops[0], MockLLIL)
-    addr_val, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-    assert isinstance(
-        addr_val, int
-    ), f"Address for RET must be an integer, got {type(addr_val)}"
-
-    # The lifting phase already expands a 16-bit return address to a 20-bit
-    # value using the current PC page, so both RET and RETF simply jump to the
-    # evaluated address here.
-    regs.set(RegisterName.PC, addr_val)
-    return None, None
-
-
-def eval_jump(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    assert isinstance(llil.ops[0], MockLLIL)
-    addr, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-    assert isinstance(
-        addr, int
-    ), f"Address for JUMP must be an integer, got {type(addr)}"
-    regs.set(RegisterName.PC, addr)
-    return None, None
-
-
-def eval_call(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    assert isinstance(llil.ops[0], MockLLIL)
-    addr, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-    assert isinstance(
-        addr, int
-    ), f"Address for CALL must be an integer, got {type(addr)}"
-
-    ret_addr = regs.get(RegisterName.PC)
-
-    # Determine push size for return address
-    # SC62015: CALL mn (2-byte target) pushes 2 bytes; CALLF lmn (3-byte target) pushes 3 bytes.
-    # This distinction should ideally be made during lifting.
-    # If llil.ops[0] (target addr expr) resulted from a 16-bit const, it's likely a short CALL.
-    push_size = 3  # Default for CALLF / system stack operations
-    if (
-        llil.ops[0].op == "CONST_PTR.w"
-        or (llil.ops[0].op == "CONST.w")
-        or (llil.ops[0].op == "OR.l" and llil.ops[0].ops[0].op == "CONST.w")
-    ):  # Heuristic for CALL mn
-        push_size = 2
-
-    stack_addr = regs.get(RegisterName.S) - push_size
-    # Ensure ret_addr is masked correctly for the push size
-    if push_size == 2:
-        memory.write_bytes(push_size, stack_addr, ret_addr & 0xFFFF)
-    else:  # push_size == 3
-        memory.write_bytes(push_size, stack_addr, ret_addr & 0xFFFFF)
-
-    regs.set(RegisterName.S, stack_addr)
-    regs.set(RegisterName.PC, addr)
-    return None, None
-
-
-
-
-def eval_cmp_e(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[int, Optional[ResultFlags]]:
-    assert isinstance(llil.ops[0], MockLLIL) and isinstance(llil.ops[1], MockLLIL)
-    op1_val, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-    op2_val, _ = evaluate_llil(llil.ops[1], regs, memory, state)
-    assert op1_val is not None and op2_val is not None
-    return int(int(op1_val) == int(op2_val)), None
-
-
-def eval_cmp_ugt(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[int, Optional[ResultFlags]]:
-    assert isinstance(llil.ops[0], MockLLIL) and isinstance(llil.ops[1], MockLLIL)
-    op1_val, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-    op2_val, _ = evaluate_llil(llil.ops[1], regs, memory, state)
-    assert op1_val is not None and op2_val is not None
-    return int(int(op1_val) > int(op2_val)), None
-
-
-def to_signed(value: int, size_bytes: int) -> int:
-    width_bits = size_bytes * 8
-    mask = (1 << width_bits) - 1
-    sign_bit_mask = 1 << (width_bits - 1)
-    value &= mask
-    if (value & sign_bit_mask) != 0:
-        return value - (1 << width_bits)
-    return value
-
-
-def eval_cmp_slt(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[int, Optional[ResultFlags]]:
-    assert size is not None, "Size must be provided for signed comparison"
-    assert isinstance(llil.ops[0], MockLLIL) and isinstance(llil.ops[1], MockLLIL)
-    op1_val, _ = evaluate_llil(llil.ops[0], regs, memory, state)
-    op2_val, _ = evaluate_llil(llil.ops[1], regs, memory, state)
-    assert op1_val is not None and op2_val is not None
-
-    signed_op1 = to_signed(int(op1_val), size)
-    signed_op2 = to_signed(int(op2_val), size)
-
-    return int(signed_op1 < signed_op2), None
-
-
-def _lsl_impl(size: int, val: int, count: int) -> Tuple[int, ResultFlags]:
-    width = size * 8
-    mask = (1 << width) - 1
-    if count == 0:
-        arith_result = val & mask
-        return arith_result, {"C": 0, "Z": 1 if arith_result == 0 else 0}
-
-    carry_out = 0
-    if count <= width and width > 0:
-        carry_out = (val >> (width - count)) & 1
-
-    arith_result = (val << count) & mask
-    zero_flag = 1 if arith_result == 0 else 0
-
-    return arith_result, {"C": carry_out, "Z": zero_flag}
-
-
-def _lsr_impl(size: int, val: int, count: int) -> Tuple[int, ResultFlags]:
-    width = size * 8
-    if count == 0:
-        arith_result = val & ((1 << width) - 1 if width > 0 else 0)
-        return arith_result, {"C": 0, "Z": 1 if arith_result == 0 else 0}
-
-    carry_out = 0
-    if count > 0 and count <= width and width > 0:
-        carry_out = (val >> (count - 1)) & 1
-
-    arith_result = val >> count
-    zero_flag_val = arith_result & ((1 << width) - 1 if width > 0 else 0)
-    zero_flag = 1 if zero_flag_val == 0 else 0
-
-    return arith_result, {"C": carry_out, "Z": zero_flag}
-
-
-def _ror_impl(size: int, val: int, count: int) -> Tuple[int, ResultFlags]:
-    width = size * 8
-    mask = (1 << width) - 1 if width > 0 else 0
-    if width == 0:
-        return val & mask, {"C": 0, "Z": 1 if (val & mask) == 0 else 0}
-
-    count %= width
-    if count == 0:
-        arith_result = val & mask
-        return arith_result, {"C": val & 1, "Z": 1 if arith_result == 0 else 0}
-
-    shifted_part = val >> count
-    rotated_part = val << (width - count)
-    arith_result = (shifted_part | rotated_part) & mask
-
-    carry_out = (val >> (count - 1)) & 1
-    zero_flag = 1 if arith_result == 0 else 0
-
-    return arith_result, {"C": carry_out, "Z": zero_flag}
-
-
-def _rol_impl(size: int, val: int, count: int) -> Tuple[int, ResultFlags]:
-    width = size * 8
-    mask = (1 << width) - 1 if width > 0 else 0
-    if width == 0:
-        return val & mask, {"C": 0, "Z": 1 if (val & mask) == 0 else 0}
-
-    count %= width
-    if count == 0:
-        arith_result = val & mask
-        return arith_result, {
-            "C": (val >> (width - 1)) & 1 if width > 0 else 0,
-            "Z": 1 if arith_result == 0 else 0,
-        }
-
-    shifted_part = val << count
-    rotated_part = val >> (width - count)
-    arith_result = (shifted_part | rotated_part) & mask
-
-    carry_out = (val >> (width - count)) & 1
-    zero_flag = 1 if arith_result == 0 else 0
-
-    return arith_result, {"C": carry_out, "Z": zero_flag}
-
-
-def _rrc_impl(
-    size: int, val: int, count: int, carry_in: int
-) -> Tuple[int, ResultFlags]:
-    width = size * 8
-    mask = (1 << width) - 1 if width > 0 else 0
-    assert count == 1, "RRC count should be 1 for standard definition"
-
-    if width == 0:
-        return val & mask, {"C": 0, "Z": 1 if (val & mask) == 0 else 0}
-
-    new_carry_out = val & 1
-    arith_result = (val >> count) | (carry_in << (width - count))
-    arith_result &= mask
-
-    zero_flag = 1 if arith_result == 0 else 0
-    return arith_result, {"C": new_carry_out, "Z": zero_flag}
-
-
-def _rlc_impl(
-    size: int, val: int, count: int, carry_in: int
-) -> Tuple[int, ResultFlags]:
-    width = size * 8
-    mask = (1 << width) - 1 if width > 0 else 0
-    assert count == 1, "RLC count should be 1 for standard definition"
-
-    if width == 0:
-        return val & mask, {"C": 0, "Z": 1 if (val & mask) == 0 else 0}
-
-    new_carry_out = (val >> (width - 1)) & 1 if width > 0 else 0
-    arith_result = (val << count) | carry_in
-    arith_result &= mask
-
-    zero_flag = 1 if arith_result == 0 else 0
-    return arith_result, {"C": new_carry_out, "Z": zero_flag}
-
-
-
-
-def eval_intrinsic_tcl(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    return None, None
-
-
-def eval_intrinsic_halt(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    state.halted = True
-    return None, None
-
-
-def eval_intrinsic_off(
-    llil: MockLLIL, size: Optional[int], regs: Registers, memory: Memory, state: State
-) -> Tuple[None, Optional[ResultFlags]]:
-    state.halted = True
-    return None, None
-
-
-EVAL_LLIL: Dict[str, EvalLLILType] = {
-    "CONST": eval_const,
-    "CONST_PTR": eval_const_ptr,
-    "REG": eval_reg,
-    "SET_REG": eval_set_reg,
-    "FLAG": eval_flag,
-    "SET_FLAG": eval_set_flag,
-    "AND": _create_logical_eval(lambda a, b: a & b),
-    "OR": _create_logical_eval(lambda a, b: a | b),
-    "XOR": _create_logical_eval(lambda a, b: a ^ b),
-    "POP": eval_pop,
-    "PUSH": eval_push,
-    "NOP": eval_nop,
-    "UNIMPL": eval_unimpl,
-    "STORE": eval_store,
-    "LOAD": eval_load,
-    "RET": eval_ret,
-    "JUMP": eval_jump,
-    "CALL": eval_call,
-    "ADD": _create_arithmetic_eval(
-        lambda a, b: a + b,
-        lambda _a, _b, result, mask: 1 if result > mask else 0,
-    ),
-    "SUB": _create_arithmetic_eval(
-        lambda a, b: a - b,
-        lambda _a, _b, result, _mask: 1 if result < 0 else 0,
-    ),
-    "CMP_E": eval_cmp_e,
-    "CMP_UGT": eval_cmp_ugt,
-    "CMP_SLT": eval_cmp_slt,
-    "LSL": _create_shift_eval(lambda size, val, count: _lsl_impl(size, val, count)),
-    "LSR": _create_shift_eval(lambda size, val, count: _lsr_impl(size, val, count)),
-    "ROR": _create_shift_eval(lambda size, val, count: _ror_impl(size, val, count)),
-    "RRC": _create_shift_eval(
-        lambda size, val, count, carry: _rrc_impl(size, val, count, carry)
-    ),
-    "ROL": _create_shift_eval(lambda size, val, count: _rol_impl(size, val, count)),
-    "RLC": _create_shift_eval(
-        lambda size, val, count, carry: _rlc_impl(size, val, count, carry)
-    ),
-    "INTRINSIC_TCL": eval_intrinsic_tcl,
-    "INTRINSIC_HALT": eval_intrinsic_halt,
-    "INTRINSIC_OFF": eval_intrinsic_off,
-}
